Rename folder to syntaxChecker
