+++ client/src/external/handlers/syntaxChecker/base.ts
