+++ client/src/external/handlers/syntaxChecker/getCLCheckerCPPSrc.ts
