+++ client/src/external/handlers/syntaxChecker/utils.ts
